requirements.txt
Flask>=2.2.0
Pillow==9.1.1
